refactor: MemberTestDataConfig와 AdminTestDataConfig 의 공통 기본 클래스 생성
